Create FALLTHROUGH macro for use in switch statements
Instead of relying on special comments, use the new [[fallthrough]]
annotation, or (failing that), the non-standard
__attribute__((fallthrough) to indicate case clauses which shouldn't
end with a break statement.

+++ newlib/libc/stdio/fvwrite.c
@@ -220,7 +220,7 @@ __sfvwrite_r (struct _reent *ptr,
 	  if (!nlknown)
 	    {
 	      nl = memchr ((void *) p, '\n', len);
-	      nldist = nl ? nl + 1 - p : len + 1;
+	      nldist = nl ? nl + 1 - p : (int) (len + 1);
 	      nlknown = 1;
 	    }
 	  s = MIN (len, nldist);
